% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accuracy}
\alias{accuracy}
\title{Get percentage of parameters that are in HPD & mean relative bias from simdata}
\usage{
accuracy(smry)
}
\arguments{
\item{smry}{fastan summary}
}
\description{
Get percentage of parameters that are in HPD & mean relative bias from simdata
}
