% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{adjust_data_interface}
\alias{adjust_data_interface}
\title{Adjust the data argument on \code{rstan::stan()}}
\usage{
adjust_data_interface(model)
}
\arguments{
\item{model}{fastan model object.}
}
\value{
list; goes on the data argument in \code{rstan::stan()}.
}
\description{
Adjust the data argument on \code{rstan::stan()}
}
