% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{alpha_cov_to_var}
\alias{alpha_cov_to_var}
\title{Create matrix var(alpha) from covariances on prior}
\usage{
alpha_cov_to_var(prior)
}
\arguments{
\item{prior}{.}
}
\description{
Create matrix var(alpha) from covariances on prior
}
