% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Transform summary matrix into data frame}
\usage{
as_data_frame(smry, par, row = NULL, col = NULL, stat = c("mean", "median"))
}
\arguments{
\item{smry}{.}

\item{par}{.}

\item{row}{.}

\item{col}{.}

\item{stat}{.}
}
\description{
Transform summary matrix into data frame
}
