% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_engines.R
\name{car}
\alias{car}
\title{CAR operation}
\usage{
car(W, rho = 0.95)
}
\arguments{
\item{W}{matrix, neighborhood matrix.}

\item{rho}{numeric, default = 0.95, car parameter.}
}
\value{
covariance matrix.
}
\description{
CAR operation
}
