% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_engines.R
\name{car_conditional}
\alias{car_conditional}
\title{prior engine: CAR conditional for alpha covariance in models with more than one group}
\usage{
car_conditional(proj, neib, tau)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{neib}{function that calculates the neighborhood matrix.}

\item{tau}{numeric, to scale CAR matrix.}
}
\value{
list of covariance matrices.
}
\description{
prior engine: CAR conditional for alpha covariance in models with more than one group
}
