% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_engines.R
\name{car_engine}
\alias{car_engine}
\title{Prior engine for CAR}
\usage{
car_engine(proj, type, neib = NULL, tau = 1)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{type}{string, type of CAR, can be either 'conditional', 'expl', or 'simple'.}

\item{neib}{function that calculates the neighborhood matrix (default NULL, for when type == "siple").}

\item{tau}{numeric, to scale CAR matrix, default is 1.}
}
\value{
list of covariance matrices.
}
\description{
Prior engine for CAR
}
