% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_engines.R
\name{car_simple}
\alias{car_simple}
\title{prior engine: CAR with simple neighborhood matrix}
\usage{
car_simple(proj, tau)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{tau}{numeric, to scale CAR matrix.}
}
\value{
list of covariance matrices.
}
\description{
(used only for time dependency on lambdas)
}
