% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{delete}
\alias{delete}
\title{Remove attributes from \code{fastan::project}}
\usage{
delete(proj, ...)
}
\arguments{
\item{proj}{fastan project object}

\item{...}{strings, attributes to be removed}
}
\description{
Remove attributes from \code{fastan::project}
}
