% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{diagnostic}
\alias{diagnostic}
\title{Get diagnostic statistics from \code{rstan::stanfit}}
\usage{
diagnostic(fit)
}
\arguments{
\item{fit}{\code{rstan::stanfit} object.}
}
\value{
\code{tibble::tibble} containt diagnostic statistics.
}
\description{
Get diagnostic statistics from \code{rstan::stanfit}
}
