% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{distances}
\alias{distances}
\title{Get distances (in km) between points}
\usage{
distances(coor, lon = "lon", lat = "lat", pairs = F)
}
\arguments{
\item{coor}{data.frame with coordinates.}

\item{lon}{string, default = "lon", column with longitude.}

\item{lat}{string, default = "lat", column with latitude.}

\item{pairs}{logical, default = \code{FALSE}, if \code{TRUE} returns data.frame with distances between each pair of point, if \code{FALSE} returns distance matrix}
}
\value{
matrix/data.frame with distances in km.
}
\description{
(to be used in project space)
}
