% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elapsed_time_table}
\alias{elapsed_time_table}
\title{Get table with elapsed time of \code{stan::stanfit} in hours}
\usage{
elapsed_time_table(fit)
}
\arguments{
\item{fit}{\code{rstan::stanfit} object.}
}
\description{
(with margins)
}
