% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{fiat_init}
\alias{fiat_init}
\title{FALTA PREDICTION}
\usage{
fiat_init(model, chains)
}
\arguments{
\item{model}{fastanModel object.}

\item{chains}{integer; number of chains.}
}
\value{
list; for each
}
\description{
Generate initial values for STAN MCMC
}
