% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{fiat_init_from_fit}
\alias{fiat_init_from_fit}
\title{Generate initial values for STAN MCMC from another fit}
\usage{
fiat_init_from_fit(draws)
}
\arguments{
\item{fit}{fastanModel object.}
}
\value{
list; for each
}
\description{
Generate initial values for STAN MCMC from another fit
}
