% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{fiat_init_from_last_value}
\alias{fiat_init_from_last_value}
\title{Generate initial values for STAN MCMC from another fit}
\usage{
fiat_init_from_last_value(fit)
}
\arguments{
\item{fit}{stanfit object.}
}
\value{
list; for each
}
\description{
Generate initial values for STAN MCMC from another fit
}
