% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fiat_sentinel}
\alias{fiat_sentinel}
\title{Generate sentinel values for your data}
\usage{
fiat_sentinel(m)
}
\arguments{
\item{m}{A scalar/vector/matrix of numeric values.}
}
\value{
A integer.
}
\description{
Generate distinct sentinel values based on your data (generally the first multiple of 1000 one order above the entries of your argument).
}
