% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{generate_data}
\alias{generate_data}
\title{Generate simulated data}
\usage{
generate_data(real, cicles = 1)
}
\arguments{
\item{real}{list containing real values of the parameters and group sizes.}

\item{cicles}{integer, default 1, number of cicles.}
}
\value{
\code{fasta::data} object.
}
\description{
(exploratory, confirmatory or semi-confirmatory models)
}
