% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{generate_data_sc}
\alias{generate_data_sc}
\title{Generate data (confirmatory or semi-confirmatory models)}
\usage{
generate_data_sc(rows.by.group, columns, cicles = 1, semi.conf = F)
}
\arguments{
\item{rows.by.group}{integer vector; size (number of loadings - or rows in the wider format) of each group.}

\item{columns}{integer; number of columns in the wider format.}

\item{cicles}{integer; UNDER DEVELOPMENT.}

\item{semi.conf}{boolean:
\itemize{
\item \code{TRUE} if it's semi-confirmatory (then the last group is considered to be the semi-confirmatory group);
\item \code{FALSE} if it's confirmatory (DEFAULT).
}}
}
\value{
fastan model object
}
\description{
Generate data (confirmatory or semi-confirmatory models)
}
