% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{generate_parameters_from_uniform}
\alias{generate_parameters_from_uniform}
\title{Title}
\usage{
generate_parameters_from_uniform(rows.by.group, columns, semi.conf, param)
}
\arguments{
\item{rows.by.group}{integer vector; size (number of loadings - or rows in the wider format) of each group.}

\item{columns}{integer; number of columns in the wider format.}

\item{semi.conf}{boolean:}

\item{param}{.}
}
\description{
Title
}
