% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{get_all_chains}
\alias{get_all_chains}
\title{Generate initial values for STAN MCMC from another fit}
\usage{
get_all_chains(fit)
}
\arguments{
\item{fit}{fastanModel object.}
}
\description{
Generate initial values for STAN MCMC from another fit
}
