% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{get_nearest_row}
\alias{get_nearest_row}
\title{Use LSAP algorithm to select nearest points}
\usage{
get_nearest_row(small, big, id, lon = "lon", lat = "lat")
}
\arguments{
\item{small}{data.frame (with rows without missings).}

\item{big}{data.frame (with rows with missings).}

\item{id}{string with id column.}

\item{lon}{string with longitude column.}

\item{lat}{string with latitude column.}
}
\value{
data.frame with relations by id.
}
\description{
Use LSAP algorithm to select nearest points
}
