% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_limits}
\alias{group_limits}
\title{Create list with row limit of each group}
\usage{
group_limits(x)
}
\arguments{
\item{x}{integer vector.}
}
\value{
list of limits (\verb{[[1]]} for inferior, \verb{[[2]]} for superior) of each group.
}
\description{
(to be used with \code{data$group.sizes})
}
