% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{header_col}
\alias{header_col}
\title{Shiny column with HTML header}
\usage{
header_col(title, color, height, width)
}
\arguments{
\item{title}{.}

\item{color}{.}

\item{height}{.}

\item{width}{.}
}
\description{
Shiny column with HTML header
}
