% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{init}
\alias{init}
\title{Generate constant initial values for STAN MCMC}
\usage{
init(proj, chains)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{chains}{integer, number of chains.}
}
\value{
list compatible with STAN init.
}
\description{
Generate constant initial values for STAN MCMC
}
