% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{init_from_fit}
\alias{init_from_fit}
\title{Generate initial values for STAN MCMC from another fit}
\usage{
init_from_fit(fit)
}
\arguments{
\item{fit}{stanfit object.}
}
\value{
list; for each
}
\description{
Generate initial values for STAN MCMC from another fit
}
