% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{interface_normal}
\alias{interface_normal}
\title{To be used with \code{fastan::interface}}
\usage{
interface_normal(proj)
}
\arguments{
\item{proj}{\code{fastan::project} object.}
}
\description{
To be used with \code{fastan::interface}
}
