% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{invert_lambda_signal}
\alias{invert_lambda_signal}
\title{Invert the signal of a column of lambda on the MCMC}
\usage{
invert_lambda_signal()
}
\description{
Invert the signal of a column of lambda on the MCMC
}
