% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{invert_signal}
\alias{invert_signal}
\title{Inver signal in factor}
\usage{
invert_signal(proj, fac, bias.stat = "mean")
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{fac}{integer vector with number the of factors to invert.}

\item{bias.stat}{sting, default = "mean", Bayes estimator to use to calculate bias, if data is simdata.}
}
\description{
Inver signal in factor
}
