% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{invert_signal_smry}
\alias{invert_signal_smry}
\title{Title}
\usage{
invert_signal_smry(smry, fac, bias.stat = "mean")
}
\arguments{
\item{smry}{.}

\item{fac}{.}

\item{bias.stat}{.}
}
\description{
Title
}
