% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Transform 3D matrix into elongated data frame}
\usage{
matrix_to_df(m)
}
\arguments{
\item{m}{matrix 3d, the 3rd dim will be elongated.}
}
\value{
\code{data.frame()}
}
\description{
Transform 3D matrix into elongated data frame
}
