% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{merge_draws}
\alias{merge_draws}
\title{Generate initial values for STAN MCMC from another fit}
\usage{
merge_draws(draws1, draws2)
}
\arguments{
\item{draws1, draws2}{draws}
}
\description{
Generate initial values for STAN MCMC from another fit
}
