% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{missing_validation}
\alias{missing_validation}
\title{Missing validation}
\usage{
missing_validation(proj, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{...}{dot args.}
}
\description{
Missing validation
}
