% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{missing_validation_selection}
\alias{missing_validation_selection}
\title{Missing validation algorithm}
\usage{
missing_validation_selection(proj)
}
\arguments{
\item{proj}{\code{fastan::project} object.}
}
\value{
\code{fastan::data} object.
}
\description{
Missing validation algorithm
}
