% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{model_data_cluster}
\alias{model_data_cluster}
\title{Process data in \code{fastan} model format (cluster model)}
\usage{
model_data_cluster()
}
\description{
Process data in \code{fastan} model format (cluster model)
}
