% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{model_data_sc}
\alias{model_data_sc}
\title{Process data in \code{fastan} model format (confirmatory or semi-confirmatory models)}
\usage{
model_data_sc(data, value, group, row, col, semi.conf)
}
\arguments{
\item{data}{data.frame in the longer format.}

\item{value}{string; name of the column containing values.}

\item{group}{string; name of the column containing groups.}

\item{row}{string; name of the column containing rows/loadings.}

\item{col}{string; name of the column containing columns/factor-levels.}
}
\value{
\code{fastan} model object
About sentinel
\itemize{
\item sentinel + 0 missing data for mcmc
\item sentinel + 1 missing data, but do nothing
\item sentinel + 2 data for test of adjustment
}
}
\description{
Process data in \code{fastan} model format (confirmatory or semi-confirmatory models)
}
