% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n.fac}
\alias{n.fac}
\title{Get number of factors of a model}
\usage{
n.fac(proj)
}
\arguments{
\item{proj}{\code{fastan::project} object.}
}
\value{
A integer, number of factor in the model.
}
\description{
Get number of factors of a model
}
