% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{neib_simple}
\alias{neib_simple}
\title{Generate simple neighborhood matrix for line spaces.}
\usage{
neib_simple(n)
}
\arguments{
\item{n}{integer, number of elements in line.}
}
\value{
neighborhood matrix.
}
\description{
Generate simple neighborhood matrix for line spaces.
}
