% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{neib_voronoi}
\alias{neib_voronoi}
\title{Calculate neighborhood matrix from Voronoi tessellation}
\usage{
neib_voronoi(coor, lon = "lon", lat = "lat")
}
\arguments{
\item{coor}{data.frame with coordinates.}

\item{lon}{string, default = "lon", column with longitude.}

\item{lat}{string, default = "lat", column with latitude.}
}
\value{
neighborhood matrix.
}
\description{
Calculate neighborhood matrix from Voronoi tessellation
}
