% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{percentage_hits}
\alias{percentage_hits}
\title{Get percentage of parameters that are in HPD from simdata}
\usage{
percentage_hits(smry)
}
\arguments{
\item{smry}{fastan summary}
}
\description{
Get percentage of parameters that are in HPD from simdata
}
