% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all_diagnostic}
\alias{plot_all_diagnostic}
\title{Plot diagnostic stats in grid for each parameter}
\usage{
plot_all_diagnostic(fit, stat)
}
\arguments{
\item{fit}{.}

\item{stat}{.}
}
\description{
Plot diagnostic stats in grid for each parameter
}
