% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bias}
\alias{plot_bias}
\title{Plot bias of simdata}
\usage{
plot_bias(smry, par = "all", correct = T)
}
\arguments{
\item{smry}{\code{fastan::summary} object.}

\item{par}{string, parameter name.}

\item{correct}{.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot bias of simdata
}
