% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_contrast}
\alias{plot_contrast}
\title{Plot contrast matrix}
\usage{
plot_contrast(smry, par = "alpha", stat = "mean")
}
\arguments{
\item{smry}{\code{fastan::summary} object.}

\item{par}{string, parameter name.}

\item{stat}{string.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot contrast matrix
}
