% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_everything}
\alias{plot_everything}
\title{Export most common plots of a project}
\usage{
plot_everything(proj)
}
\arguments{
\item{proj}{fastan project}
}
\description{
Export most common plots of a project
}
