% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hpd}
\alias{plot_hpd}
\title{Plot HPD interval & other stats}
\usage{
plot_hpd(
  proj,
  par,
  row = NULL,
  col = NULL,
  stat = "mean",
  omit.alpha0 = T,
  omit.alpha0.list = F
)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{par}{string, parameter name.}

\item{row}{integer, row number in parameter matrix.}

\item{col}{integer, column number in parameter matrix.}

\item{stat}{string.}

\item{omit.alpha0}{logical, \code{TRUE} if the zeros in prior$alpha$in_group should be omitted, \code{FALSE} otherwise.}

\item{omit.alpha0.list}{logical, default = \code{FALSE}, \code{TRUE} if just the list of ggplots should be returned, \code{FALSE} if the gridExtra plot, (makes sense only when omit.alpha0 = \code{TRUE}, used only for the shiny app).}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot HPD interval & other stats
}
