% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lambda}
\alias{plot_lambda}
\title{Plot lambdas}
\usage{
plot_lambda(smry, stat = "mean")
}
\arguments{
\item{smry}{\code{fastan::summary} object.}

\item{stat}{string.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot lambdas
}
