% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_map}
\alias{plot_map}
\title{Plot basic map structure}
\usage{
plot_map(proj)
}
\arguments{
\item{proj}{\code{fastan::project}}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot basic map structure
}
