% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_map_data}
\alias{plot_map_data}
\title{Plot map data}
\usage{
plot_map_data(proj, stat)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{stat}{string.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot map data
}
