% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_map_post}
\alias{plot_map_post}
\title{Plot map posterior statistics}
\usage{
plot_map_post(proj, par, col = 1, stat)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{par}{string, parameter name.}

\item{col}{integer, column number in parameter matrix.}

\item{stat}{string.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot map posterior statistics
}
