% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot missing values pattern}
\usage{
plot_missing(data)
}
\arguments{
\item{data}{\code{fastan::data} object.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot missing values pattern
}
