% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mock_doc}
\alias{plot_mock_doc}
\title{Mock function for the sole purpose of documentation}
\usage{
plot_mock_doc(proj, fit, smry, data, par, row, col, stat)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{fit}{\code{rstan::stanfit} object.}

\item{smry}{\code{fastan::summary} object.}

\item{data}{\code{fastan::data} object.}

\item{par}{string, parameter name.}

\item{row}{integer, row number in parameter matrix.}

\item{col}{integer, column number in parameter matrix.}

\item{stat}{string.}
}
\description{
Mock function for the sole purpose of documentation
}
