% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior sample histogram and/or density}
\usage{
plot_posterior(fit, par, row = 1, col = 1, type = c("hist", "dens"))
}
\arguments{
\item{fit}{\code{rstan::stanfit} object.}

\item{par}{string, parameter name.}

\item{row}{integer, row number in parameter matrix.}

\item{col}{integer, column number in parameter matrix.}

\item{type}{string, "hist" for just histogram, "dens" for only density, default is both.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot posterior sample histogram and/or density
}
