% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot Posterior sample trace plot}
\usage{
plot_trace(fit, par, row = 1, col = 1)
}
\arguments{
\item{fit}{\code{rstan::stanfit} object.}

\item{par}{string, parameter name.}

\item{row}{integer, row number in parameter matrix.}

\item{col}{integer, column number in parameter matrix.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot Posterior sample trace plot
}
