% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print_table}
\alias{print_table}
\title{Print LaTeX table}
\usage{
print_table(smry, par, row = NULL, col = NULL, stat = c("mean", "median"))
}
\arguments{
\item{smry}{.}

\item{par}{.}

\item{row}{.}

\item{col}{.}

\item{stat}{.}
}
\description{
Print LaTeX table
}
