% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{prior_normal}
\alias{prior_normal}
\title{Generate prior with independence between parameters}
\usage{
prior_normal(data, semi.conf, omit.alpha0 = T, ...)
}
\arguments{
\item{data}{\code{fastan::data} object.}

\item{semi.conf}{logical, \code{TRUE} if model is semi-confirmatory, \code{FALSE} otherwise.}

\item{omit.alpha0}{logical, default = \code{TRUE}, \code{TRUE} if you want to force a constant prior for alphas not in group, \code{FALSE} otherwise.}

\item{...}{dot args.}
}
\value{
\code{fastan::prior} object.
}
\description{
Generate prior with independence between parameters
}
