% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{project}
\alias{project}
\title{Generate new \code{fastan::project}}
\usage{
project(proj = NULL)
}
\arguments{
\item{proj}{list to be turned into project, if NULL (default) a blank new project is returned.}
}
\description{
Generate new \code{fastan::project}
}
