% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop.missing}
\alias{prop.missing}
\title{Calculate proportion of missing values in data}
\usage{
prop.missing(data)
}
\arguments{
\item{data}{\code{fastan::data} object.}
}
\value{
a number between 0 and 1.
}
\description{
Calculate proportion of missing values in data
}
