% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{real_from_dist}
\alias{real_from_dist}
\title{Title}
\usage{
real_from_dist(group.sizes, columns, semi.conf, dist = list())
}
\arguments{
\item{group.sizes}{.}

\item{columns}{.}

\item{semi.conf}{.}

\item{dist}{.}
}
\description{
Title
}
