% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{real_from_posterior}
\alias{real_from_posterior}
\title{Generate real value list from posterior}
\usage{
real_from_posterior(proj, stat = "mean")
}
\arguments{
\item{proj}{\code{fastan::project}.}

\item{stat}{string with pontual Bayes estimator on project summary.}
}
\value{
real value list.
}
\description{
Used on \code{fastan::generate_data}
}
