% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{real_from_prior}
\alias{real_from_prior}
\title{Generate real value list from prior}
\usage{
real_from_prior(proj)
}
\arguments{
\item{proj}{\code{fastan::project}.}
}
\value{
real value list.
}
\description{
Used on \code{fastan::generate_data}
}
