% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{real_from_uniform}
\alias{real_from_uniform}
\title{Title}
\usage{
real_from_uniform(
  group.sizes,
  columns,
  semi.conf,
  param = list(alpha = c(1, 6), lambda = c(0.5, 1.5), sigma2 = c(1, 3))
)
}
\arguments{
\item{group.sizes}{.}

\item{columns}{.}

\item{semi.conf}{.}

\item{param}{.}
}
\description{
Title
}
