% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{remove}
\alias{remove}
\title{Remove attributes from \code{fastan::project}}
\usage{
remove(proj, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{...}{dot args.}
}
\description{
Remove attributes from \code{fastan::project}
}
