% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{run_stan}
\alias{run_stan}
\title{Run STAN MCMC}
\usage{
run_stan(model, init = NULL, chains = 1, pred = F, ...)
}
\arguments{
\item{model}{.}

\item{init}{.}

\item{chains}{.}

\item{...}{arguments that will be passed to \code{rstan::stan()}}
}
\value{
rstan::stan fit object.
}
\description{
Run STAN MCMC
}
