% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{server}
\alias{server}
\title{Shiny server}
\usage{
server(proj = NULL, input, output, session)
}
\arguments{
\item{proj}{.}

\item{input}{.}

\item{output}{.}

\item{session}{.}
}
\description{
Shiny server
}
