% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_data}
\alias{set_data}
\title{tidy: set data in project}
\usage{
set_data(proj, simdata = "", pred = NULL, cicles = 1, seed = NULL, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{simdata}{logical/string, simdata type.}

\item{pred}{numeric/NULL, if data is simdata.}

\item{cicles}{numeric.}

\item{seed}{numeric.}

\item{...}{dot args.}
}
\description{
tidy: set data in project
}
