% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_diagnostic}
\alias{set_diagnostic}
\title{tidy: set diagnostic in project}
\usage{
set_diagnostic(proj, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{...}{dot args.}
}
\description{
tidy: set diagnostic in project
}
