% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_fit}
\alias{set_fit}
\title{tidy: set fit in project}
\usage{
set_fit(proj, set_summary = T, set_diagnostic = T, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{set_summary}{logical, \code{TRUE} (default) for try to set summary, \code{FALSE} otherwise.}

\item{set_diagnostic}{logical, \code{TRUE} (default) for try to set diagnostic, \code{FALSE} otherwise.}

\item{...}{dot args.}
}
\description{
tidy: set fit in project
}
