% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_info}
\alias{set_info}
\title{tidy: set information in project}
\usage{
set_info(proj, info)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{info}{string with info}
}
\description{
tidy: set information in project
}
