% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_prior}
\alias{set_prior}
\title{Title}
\usage{
set_prior(proj, type = "normal", semi.conf, engine = NULL, ...)

set_prior(proj, type = "normal", semi.conf, engine = NULL, ...)
}
\arguments{
\item{proj}{.}

\item{type}{.}

\item{semi.conf}{.}

\item{engine}{.}

\item{...}{.}
}
\description{
Title

Title
}
