% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_space}
\alias{set_space}
\title{tidy: set space in project}
\usage{
set_space(proj, type = "real", seed = NULL, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{type}{string, default = "real", if "random" a random set of coordinates is generated, if "real" a data.frame with coordinates is expected.}

\item{seed}{numeric, seed in case of simdata (only makes sense if type = "random"), if NULL (default) no seed is set.}

\item{...}{dot args.}
}
\description{
tidy: set space in project
}
