% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_summary}
\alias{set_summary}
\title{tidy: set summary in project}
\usage{
set_summary(proj, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{...}{dot args.}
}
\description{
tidy: set summary in project
}
