% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny}
\alias{shiny}
\title{Shiny App}
\usage{
shiny(proj = NULL, upload.size = 500)
}
\arguments{
\item{proj}{\code{fastan::project} object, optional.}

\item{upload.size}{numeric, default = 500, maximum upload size (in Mb).}
}
\value{
shinyApp
}
\description{
Shiny App
}
