% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_server}
\alias{shiny_server}
\title{Shiny server}
\usage{
shiny_server(proj = NULL, input, output, session)
}
\arguments{
\item{proj}{\code{fastan::project} object, optional.}

\item{input}{shiny input.}

\item{output}{shiny output.}

\item{session}{shiny server.}
}
\description{
Shiny server
}
