% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_process}
\alias{space_process}
\title{Title}
\usage{
space_process(data, df, label, lat, lon, alt, position = "row")
}
\arguments{
\item{data}{.}

\item{df}{.}

\item{label}{.}

\item{lat}{.}

\item{lon}{.}

\item{alt}{.}

\item{position}{.}
}
\description{
Title
}
