% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{stan}
\alias{stan}
\title{Run STAN MCMC}
\usage{
stan(proj, init = NULL, chains = 1, pred = T, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{init}{initial value, default = NULL, if NULL \code{fastan::init} is called.}

\item{chains}{integer, default = 1, number of MCMC chains.}

\item{pred}{logical, default = \code{TRUE}, \code{TRUE} if the predictions should be made, \code{FALSE} if otherwise.}

\item{...}{arguments that will be passed to \code{rstan::stan}.}
}
\value{
\code{rstan::stanfit} object.
}
\description{
Run STAN MCMC
}
