% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{summary_as_df}
\alias{summary_as_df}
\title{Transform summary from list of 3D matrices to list of tibbles}
\usage{
summary_as_df(proj, par = NULL)
}
\arguments{
\item{proj}{\code{fastan::project} object}

\item{par}{string vector, parameters to be returnd, default = NULL returns all parameters.}
}
\value{
list of \code{tibble:tibble}.
}
\description{
Transform summary from list of 3D matrices to list of tibbles
}
