% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{summary_matrix}
\alias{summary_matrix}
\title{Get summary from \code{rstan::stan()} fit object}
\usage{
summary_matrix(proj, bias.stat = "mean")
}
\arguments{
\item{proj}{.}

\item{bias.stat}{.}
}
\value{
list of matrices.
}
\description{
Get summary from \code{rstan::stan()} fit object
}
