% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{to_data}
\alias{to_data}
\title{Convert a object to fastan data format}
\usage{
to_data(obj)
}
\arguments{
\item{obj}{object to be turned into class in question.}
}
\description{
Convert a object to fastan data format
}
