% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{to_prior}
\alias{to_prior}
\title{Convert a object to fastan prior format}
\usage{
to_prior(obj)
}
\arguments{
\item{obj}{object to be turned into class in question.}
}
\description{
Convert a object to fastan prior format
}
