% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{to_summary}
\alias{to_summary}
\title{Convert a object to fastan summary format}
\usage{
to_summary(obj)
}
\arguments{
\item{obj}{object to be turned into class in question.}
}
\description{
Convert a object to fastan summary format
}
