% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_proj_arg}
\alias{validate_proj_arg}
\title{Validates argument in functions that accepct either the whole \code{fastan::project} or one of its attributes}
\usage{
validate_proj_arg(obj, class)
}
\arguments{
\item{obj}{\code{fatan::project} object.}

\item{class}{string, \code{fatan::project} attribute.}
}
\description{
Validates argument in functions that accepct either the whole \code{fastan::project} or one of its attributes
}
