% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export}
\alias{export}
\title{Export most common plots of a project}
\usage{
export(proj, path_dump = getwd(), rds = T, plot.extension = "png")
}
\arguments{
\item{proj}{fastan project}

\item{path_dump}{.}

\item{rds}{.}

\item{plot.extension}{.}
}
\description{
Export most common plots of a project
}
