% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{fiat_init}
\alias{fiat_init}
\title{Generate initial values for STAN MCMC}
\usage{
fiat_init(proj, chains)
}
\arguments{
\item{proj}{fastan project object.}

\item{chains}{integer; number of chains.}
}
\value{
list; for each
}
\description{
Generate initial values for STAN MCMC
}
