% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{generate_data}
\alias{generate_data}
\title{Generate data (confirmatory or semi-confirmatory models)}
\usage{
generate_data(real, cicles = 1)
}
\arguments{
\item{real}{.}

\item{cicles}{.}

\item{pred}{.}
}
\value{
fastan model object
}
\description{
Generate data (confirmatory or semi-confirmatory models)
}
