% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{get_chains_mcmc}
\alias{get_chains_mcmc}
\title{Transform \code{rstan::stanfit} into \code{coda::mcmc.list}}
\usage{
get_chains_mcmc(fit, param)
}
\arguments{
\item{fit}{\code{rstan::stanfit} object.}

\item{param}{string, must be in the \code{rstan::stanfit} format.}
}
\value{
\code{coda::mcmc.list} object.
}
\description{
Transform \code{rstan::stanfit} into \code{coda::mcmc.list}
}
