% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{get_nearest_row}
\alias{get_nearest_row}
\title{Title}
\usage{
get_nearest_row(small, big, id, lon = "lon", lat = "lat")
}
\arguments{
\item{small}{.}

\item{big}{.}

\item{id}{.}

\item{lon}{.}

\item{lat}{.}
}
\description{
Title
}
