% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{interface_normal}
\alias{interface_normal}
\title{Adjust the data argument on \code{rstan::stan()}}
\usage{
interface_normal(proj)
}
\arguments{
\item{proj}{fastan proj object.}
}
\description{
Adjust the data argument on \code{rstan::stan()}
}
