% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{model_data_sc}
\alias{model_data_sc}
\title{Process data in \code{fastan} model format (confirmatory or semi-confirmatory models)}
\usage{
model_data_sc(data, value, group, row, col, semi.conf, factor_name = NULL)
}
\arguments{
\item{data}{data.frame in the longer format.}

\item{value}{string; name of the column containing values.}

\item{group}{string; name of the column containing groups.}

\item{row}{string; name of the column containing rows/loadings.}

\item{col}{string; name of the column containing columns/factor-levels.}

\item{semi.conf}{.}

\item{factor_name}{.}
}
\value{
\code{fastan} model object
}
\description{
Process data in \code{fastan} model format (confirmatory or semi-confirmatory models)
}
