% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_contrast}
\alias{plot_contrast}
\title{Plot contrast matrix}
\usage{
plot_contrast(smry, par = "alpha", stat = "mean")
}
\arguments{
\item{smry}{.}

\item{par}{.}

\item{stat}{.}
}
\value{
ggplot2 plot
}
\description{
Plot contrast matrix
}
