% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Plot diagnostic stats}
\usage{
plot_diagnostic(proj, stat, list = F)
}
\arguments{
\item{proj}{.}

\item{stat}{.}

\item{list}{.}
}
\description{
Plot diagnostic stats
}
