% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Plot Neff or Rhat}
\usage{
plot_diagnostic(fit, stat, par = "all")
}
\arguments{
\item{fit}{.}

\item{stat}{.}

\item{par}{.}
}
\description{
Plot Neff or Rhat
}
