% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Plot diagnostic stats}
\usage{
plot_diagnostic(proj, stat, list = F)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{stat}{string.}

\item{list}{logical, default = \code{FALSE}, \code{TRUE} if just the list of ggplots should be returned, \code{FALSE} if the gridExtra plot.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
Plot diagnostic stats
}
