% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hpd}
\alias{plot_hpd}
\title{Plot HPD interval & other stats}
\usage{
plot_hpd(smry, par, row = NULL, col = NULL, stat = "mean")
}
\arguments{
\item{smry}{.}

\item{par}{.}

\item{row}{.}

\item{col}{.}

\item{stat}{.}
}
\value{
ggplot2 plot
}
\description{
Plot HPD interval & other stats
}
