% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mock_doc}
\alias{plot_mock_doc}
\title{Mock function for the sole purpose of documentation}
\usage{
plot_mock_doc(proj, fit, smry, data, par, row, col, loc, type, stat)
}
\arguments{
\item{proj}{.}

\item{fit}{.}

\item{smry}{.}

\item{data}{.}

\item{par}{.}

\item{row}{.}

\item{col}{.}

\item{loc}{.}

\item{type}{.}

\item{stat}{.}
}
\description{
Mock function for the sole purpose of documentation
}
