% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mock_doc}
\alias{plot_mock_doc}
\title{Mock function for the sole purpose of documentation}
\usage{
plot_mock_doc()
}
\arguments{
\item{fit}{.}

\item{smry}{.}

\item{mod}{.}

\item{par}{.}

\item{row}{.}

\item{col}{.}

\item{type}{.}

\item{stat}{.}

\item{warmup}{.}
}
\description{
Mock function for the sole purpose of documentation
}
