% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_normal_prior}
\alias{plot_normal_prior}
\title{Plot prior hyperparameters}
\usage{
plot_normal_prior(proj, par, stat, fac = "all")
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{par}{string, parameter name.}

\item{stat}{string.}

\item{fac}{integer, factor number.}
}
\value{
\code{ggplot2}/\code{gridExtra} object.
}
\description{
(only here for shiny app and export function)
}
