% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_normal_prior}
\alias{plot_normal_prior}
\title{Title}
\usage{
plot_normal_prior(proj, par, stat, row = NULL, col = NULL)
}
\arguments{
\item{proj}{.}

\item{par}{.}

\item{stat}{.}

\item{row}{.}

\item{col}{.}
}
\description{
Title
}
