% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior sample histogram and/or density}
\usage{
plot_posterior(fit, par, row = 1, col = 1, type = c("hist", "dens"))
}
\arguments{
\item{fit}{.}

\item{par}{.}

\item{row}{.}

\item{col}{.}

\item{type}{.}
}
\value{
ggplot2 plot
}
\description{
Plot posterior sample histogram and/or density
}
