% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{prior_normal}
\alias{prior_normal}
\title{Title}
\usage{
prior_normal(data, semi.conf, omit.alpha0 = F)
}
\arguments{
\item{data}{.}

\item{semi.conf}{.}

\item{omit.alpha0}{.}
}
\description{
Title
}
