% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{prior_normal}
\alias{prior_normal}
\title{Title}
\usage{
prior_normal(data, dependence = list(lambda = F, alpha = F), semi.conf)
}
\arguments{
\item{data}{.}

\item{dependence}{.}

\item{semi.conf}{.}
}
\description{
Title
}
