% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{real_from_dist}
\alias{real_from_dist}
\title{Title}
\usage{
real_from_dist(
  group.sizes,
  columns,
  semi.conf,
  dist = list(alpha = dist_uniform(-6, 6), lambda = dist_normal(0, 1), sigma2 =
    dist_uniform(0.1, 2))
)
}
\arguments{
\item{group.sizes}{.}

\item{columns}{.}

\item{semi.conf}{.}

\item{dist}{.}
}
\description{
Title
}
