% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{real_from_uniform}
\alias{real_from_uniform}
\title{Title}
\usage{
real_from_uniform(
  group.sizes,
  columns,
  semi.conf,
  norm = list(alpha = F, lambda = F),
  param = list(alpha = c(-6, 6), lambda = c(-1, 1), sigma2 = c(0.1, 2))
)
}
\arguments{
\item{group.sizes}{.}

\item{columns}{.}

\item{semi.conf}{.}

\item{norm}{.}

\item{param}{.}
}
\description{
Title
}
