% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_data}
\alias{set_data}
\title{Title}
\usage{
set_data(proj, simdata = "", pred = NULL, cicles = 1, seed = NULL, ...)
}
\arguments{
\item{proj}{.}

\item{simdata}{.}

\item{pred}{.}

\item{cicles}{.}

\item{seed}{.}

\item{...}{.}
}
\description{
Title
}
