% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_fit}
\alias{set_fit}
\title{Title}
\usage{
set_fit(proj, set_summary = T, set_diagnostic = T, ...)
}
\arguments{
\item{proj}{.}

\item{set_summary}{.}

\item{set_diagnostic}{.}

\item{...}{.}
}
\description{
Title
}
