% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{set_prior}
\alias{set_prior}
\title{tidy: set prior in project}
\usage{
set_prior(proj, type = "normal", semi.conf, engine = NULL, ...)
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{type}{string, prior type.}

\item{semi.conf}{logical, \code{TRUE} if model is semi-confirmatory, \code{FALSE} otherwise.}

\item{engine}{list of engines to generate prior hyperparameters, if NULL (default) then prior is independent.}

\item{...}{dot args.}
}
\description{
tidy: set prior in project
}
