% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_process}
\alias{space_process}
\title{Process spatial data associated with the project}
\usage{
space_process(data, df, label, lat, lon, alt, position = "row")
}
\arguments{
\item{data}{\code{fastan::data} object.}

\item{df}{data.frame with spatial data.}

\item{label}{string, column with id.}

\item{lat}{string, column with latitude.}

\item{lon}{string, column with longitude.}

\item{alt}{string, column with altitude.}

\item{position}{string, default = "row", if "row" data is associated with loadings, if "col" with factor scores.}
}
\value{
\code{tibble::tibble}.
}
\description{
Process spatial data associated with the project
}
