% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{stan}
\alias{stan}
\title{Run STAN MCMC}
\usage{
stan(proj, init = NULL, chains = 1, pred = T, ...)
}
\arguments{
\item{proj}{.}

\item{init}{.}

\item{chains}{.}

\item{pred}{.}

\item{...}{arguments that will be passed to \code{rstan::stan()}}
}
\value{
rstan::stanfit object.
}
\description{
Run STAN MCMC
}
