% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{summary_as_df}
\alias{summary_as_df}
\title{Transform 3D matrix into elongated data frame}
\usage{
summary_as_df(m, label = NULL)
}
\arguments{
\item{m}{matrix 3d, the 3rd dim will be elongated.}

\item{label}{.}
}
\value{
\code{data.frame()}
}
\description{
Transform 3D matrix into elongated data frame
}
