% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{summary_matrix}
\alias{summary_matrix}
\title{Get summary from \code{rstan::stan()} fit object}
\usage{
summary_matrix(fit, model = NULL)
}
\arguments{
\item{fit}{\code{rstan::stan()} fit object.}

\item{model}{fastan model object}
}
\value{
list of matrices.
}
\description{
Get summary from \code{rstan::stan()} fit object
}
