% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{summary_matrix}
\alias{summary_matrix}
\title{Get summary from \code{rstan::stanfit} object}
\usage{
summary_matrix(proj, bias.stat = "mean")
}
\arguments{
\item{proj}{\code{fastan::project} object.}

\item{bias.stat}{string, default = "mean", Bayes estimator to be used to calculate bias if data is simdata.}
}
\value{
list of matrices.
}
\description{
Get summary from \code{rstan::stanfit} object
}
